Deprecate interpret_eval naming
