+++ .gitignore
@@ -1,11 +1,11 @@
  .DS_Store
-interpret_eval/example/test-re.tsv
-interpret_eval/example/test-re.tsv
-interpret_eval/example/test_re.tsv
-interpret_eval/tasks/re/errorAnalysis.py
-interpret_eval/example/test-re.tsv
-interpret_eval/example/test-re.tsv
-interpret_eval/tasks/re/test_re.tsv
-interpret_eval/tasks/re/tensoreval-backup.py
-interpret_eval/tasks/re/a.json
-interpret_eval/tasks/re/instantiate.json
+explainaboard/example/test-re.tsv
+explainaboard/example/test-re.tsv
+explainaboard/example/test_re.tsv
+explainaboard/tasks/re/errorAnalysis.py
+explainaboard/example/test-re.tsv
+explainaboard/example/test-re.tsv
+explainaboard/tasks/re/test_re.tsv
+explainaboard/tasks/re/tensoreval-backup.py
+explainaboard/tasks/re/a.json
+explainaboard/tasks/re/instantiate.json
