+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include *.txt *.png
-recursive-include interpret_eval/tasks/* *.aspects *.json
+recursive-include explainaboard/tasks/* *.aspects *.json
 recursive-include fig *.png
