+++ explainaboard/__init__.py
@@ -1,0 +1,5 @@
+__version__ = "0.1.7"
+
+import explainaboard.explainaboard_main
+import explainaboard.tasks
+
