+++ explainaboard/tasks/nli/tensoreval.py
@@ -2,21 +2,12 @@
 import argparse
 import numpy
 import sys
+import csv
+
 # sys.path.append("./src")
 # from src.utils import *
 # from src.errorAnalysis import *
-
 from ..src.errorAnalysis import *
-
-
-
-
-
-# def tuple2str(triplet):
-# 	res = ""
-# 	for v in triplet:
-# 		res += str(v) + "_"
-# 	return res.rstrip("_")
 
 
 
@@ -26,65 +17,6 @@ def sent2list(sent):
 		return list(sent)
 	else:
 		return sent.split(" ")
-
-
-
-
-def getAspectValue(sent_list, sample_list_tag, sample_list_tag_pred, dict_aspect_func):
-
-
-
-
-
-	dict_span2aspectVal = {}
-	dict_span2aspectVal_pred = {}
-
-
-	for aspect, fun in dict_aspect_func.items():
-		dict_span2aspectVal[aspect] = {}
-		dict_span2aspectVal_pred[aspect] = {}
-
-
-	# maintain it for print error case
-	dict_sid2sent = {}
-
-
-	sample_id = 0
-	for  sent, tag, tag_pred in zip(sent_list, sample_list_tag, sample_list_tag_pred):
-
-
-		dict_sid2sent[str(sample_id)] = format4json_tc(sent)
-
-
-
-		word_list = wordSegment(sent).split(" ")
-
-
-
-		sent_length = len(word_list)
-
-		sent_pos = tuple2str((sample_id, tag))
-		sent_pos_pred = tuple2str((sample_id, tag_pred))
-
-		# Sentence Length: sentALen
-		aspect = "sLen"
-		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = float(sent_length)
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent_length)
-
-
-		# Tag: tag
-		aspect = "tag"   ############## MUST Be Gold Tag for text classification task
-		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = tag
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = tag
-
-
-
-		sample_id += 1
-
-	# print(dict_span2aspectVal["bleu"])
-	return  dict_span2aspectVal, dict_span2aspectVal_pred, dict_sid2sent
 
 
 def get_probability_right_or_not(file_path):
@@ -99,7 +31,7 @@ def get_probability_right_or_not(file_path):
     the file_path must in the format: /root/path/to/your/file/test_dataset.tsv
 
     The file must in this format:
-    sentence\tground_truth\tpredict_label\tprobability\tright_or_not
+    sentence1\tsentence2\tground_truth\tpredict_label\tprobability\tright_or_not
     if prediction is right, right_or_not is assigned to 1, otherwise 0.
 
     """
@@ -109,8 +41,8 @@ def get_probability_right_or_not(file_path):
 
     result = pd.read_csv(file_path, sep='\t', header=None)
 
-    probability_list = np.array(result[3]).tolist()
-    right_or_not_list = np.array(result[4]).tolist()
+    probability_list = np.array(result[4]).tolist()
+    right_or_not_list = np.array(result[5]).tolist()
 
     return probability_list, right_or_not_list
 
@@ -165,7 +97,7 @@ def divide_into_bin(size_of_bin, raw_list):
     for i in range(0, size_of_bin):
         value = bin_list[i]
         if len(value) == 0:
-            result_list.append([0, 0, 0])
+            result_list.append([None, None, 0])
             continue
         total_probability = 0
         total_right = 0
@@ -189,7 +121,7 @@ def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-atte
     the file_path must in the format: /root/path/to/your/file/test_dataset.tsv
 
     The file must in this format:
-    sentence\tground_truth\tpredict_label\tprobability\tright_or_not
+    sentence1\tsentence2\tground_truth\tpredict_label\tprobability\tright_or_not
     if prediction is right, right_or_not is assigned to 1, otherwise 0.
 
     :param size_of_bin: the numbers of how many bins
@@ -202,8 +134,8 @@ def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-atte
     ece :the ece of this file
     dic :the details of the ECE information in json format
     """
-
-    # import json
+    from collections import OrderedDict
+    import json
 
     probability_list, right_or_not_list = get_probability_right_or_not(file_path)
 
@@ -220,13 +152,105 @@ def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-atte
     basic_width = 1 / size_of_bin
     for i in range(len(bin_list)):
         tem_dic = {}
-
         bin_name = format(i * basic_width, '.2g') + '-' + format((i+1) * basic_width, '.2g')
         tem_dic = {'interval':bin_name,'average_accuracy': bin_list[i][1], 'average_confidence': bin_list[i][0],
                              'samples_number_in_this_bin': bin_list[i][2]}
         dic['details'].append(tem_dic)
 
     return ece, dic
+
+def getAspectValue(sent1_list, sent2_list, sample_list_tag, sample_list_tag_pred, dict_aspect_func):
+
+
+
+
+
+	dict_span2aspectVal = {}
+	dict_span2aspectVal_pred = {}
+
+	for aspect, fun in dict_aspect_func.items():
+		dict_span2aspectVal[aspect] = {}
+		dict_span2aspectVal_pred[aspect] = {}
+
+
+	# for error analysis
+	dict_sid2sentpair = {}
+
+	sample_id = 0
+	for  sent1, sent2, tag, tag_pred in zip(sent1_list, sent2_list, sample_list_tag, sample_list_tag_pred):
+
+		word_list1 = wordSegment(sent1).split(" ")
+		word_list2 = wordSegment(sent2).split(" ")
+
+		# for saving errorlist -- fine-grained version
+		dict_sid2sentpair[str(sample_id)] = format4json_tc(format4json_tc(sent1)+"|||"+format4json_tc(sent2))
+
+
+		sent1_length = len(word_list1)
+		sent2_length = len(word_list2)
+
+		sent_pos      = tuple2str((sample_id, tag))
+		sent_pos_pred = tuple2str((sample_id, tag_pred))
+
+		hypo = [wordSegment(sent2)]
+		refs = [[wordSegment(sent1)]]
+
+		# bleu = sacrebleu.corpus_bleu(hypo, refs).score * 0.01
+
+
+		# aspect = "bleu"
+		# if aspect in dict_aspect_func.keys():
+		# 	dict_span2aspectVal["bleu"][sent_pos] = bleu
+		# 	dict_span2aspectVal_pred["bleu"][sent_pos_pred] = bleu
+
+
+		# Sentence Length: sentALen
+		aspect = "sentALen"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal[aspect][sent_pos] = float(sent1_length)
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent1_length)
+
+
+		# Sentence Length: sentBLen
+		aspect = "sentBLen"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal["sentBLen"][sent_pos] = float(sent2_length)
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent2_length)
+
+
+		# The difference of sentence length: senDeltaLen
+		aspect = "A-B"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal["A-B"][sent_pos] = float(sent1_length-sent2_length)
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent1_length-sent2_length)
+
+
+		# "A+B"
+		aspect = "A+B"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal["A+B"][sent_pos] = float(sent1_length+sent2_length)
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent1_length+sent2_length)
+
+		# "A/B"
+		aspect = "A/B"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal["A/B"][sent_pos] = float(sent1_length*1.0/sent2_length)
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent1_length*1.0/sent2_length)
+
+
+		# Tag: tag
+		aspect = "tag"
+		if aspect in dict_aspect_func.keys():
+			dict_span2aspectVal["tag"][sent_pos] = tag
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = tag
+
+
+		sample_id += 1
+	# print(dict_span2aspectVal["bleu"])
+	return  dict_span2aspectVal, dict_span2aspectVal_pred, dict_sid2sentpair
+
+
+
 
 def evaluate(task_type = "ner", analysis_type = "single", systems = [], output = "./output.json", is_print_ci = False, is_print_case = False, is_print_ece = False):
 
@@ -240,9 +264,11 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 	corpus_type = "dataset_name"
 	model_name = "model_name"
 	path_preComputed = ""
-	path_aspect_conf = "./interpret_eval/tasks/tc/conf.aspects"
-	path_json_input = "./interpret_eval/tasks/tc/template.json"
+	path_aspect_conf = "./explainaboard/tasks/nli/conf.aspects"
+	path_json_input = "./explainaboard/tasks/nli/template.json"
 	fn_write_json = output
+
+
 
 
 	# Initalization
@@ -252,8 +278,6 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 	print(dict_aspect_func)
 
 	fwrite_json = open(fn_write_json, 'w')
-
-	path_comb_output = model_name + "/" + path_text.split("/")[-1]
 
 
 
@@ -268,34 +292,25 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 
 
 
-	sent_list, true_label_list, pred_label_list = file_to_list_tc(path_text)
+	sent1_list, sent2_list, true_label_list, pred_label_list = file_to_list_nli(path_text)
+
 
 	errorCase_list = []
 	if is_print_case:
-		errorCase_list = getErrorCase_tc(sent_list, true_label_list, pred_label_list)
+		errorCase_list = getErrorCase_nli(sent1_list, sent2_list, true_label_list, pred_label_list)
 		print(" -*-*-*- the number of error casse:\t", len(errorCase_list))
-
-
-
-
-
-
-
-
-
 
 
 
 	# Confidence Interval of Holistic Performance
 	confidence_low, confidence_up = 0,0
 	if is_print_ci:
-		confidence_low, confidence_up = compute_confidence_interval_acc(true_label_list, pred_label_list, n_times=1000)
+		confidence_low, confidence_up = compute_confidence_interval_acc(true_label_list, pred_label_list, n_times=100)
 
 
 
 
-	dict_span2aspectVal, dict_span2aspectVal_pred,  dict_sid2sent  = getAspectValue(sent_list, true_label_list, pred_label_list, dict_aspect_func)
-
+	dict_span2aspectVal, dict_span2aspectVal_pred, dict_sid2sentpair = getAspectValue(sent1_list, sent2_list, true_label_list, pred_label_list, dict_aspect_func)
 
 	holistic_performance = accuracy(true_label_list, pred_label_list)
 	holistic_performance = format(holistic_performance, '.3g')
@@ -339,7 +354,7 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 		dict_bucket2span_pred[aspect] = bucketAttribute_SpecifiedBucketInterval(dict_span2aspectVal_pred[aspect],
 																				dict_bucket2span[aspect].keys())
 		# dict_bucket2span_pred[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspectVal_pred[aspect])
-		dict_bucket2f1[aspect] = getBucketAcc_with_errorCase(dict_bucket2span[aspect], dict_bucket2span_pred[aspect], dict_sid2sent, is_print_ci, is_print_case)
+		dict_bucket2f1[aspect] = getBucketAcc_with_errorCase_nli(dict_bucket2span[aspect], dict_bucket2span_pred[aspect], dict_sid2sentpair, is_print_ci, is_print_case)
 		aspect_names.append(aspect)
 	print("aspect_names: ", aspect_names)
 
@@ -397,6 +412,8 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 			n_sample = v[1]
 			confidence_low = format(v[2], '.4g')
 			confidence_up  = format(v[3], '.4g')
+
+			# for saving errorlist -- finegrained version
 			bucket_error_case = v[4]
 
 			# instantiation
@@ -412,16 +429,22 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 	obj_json["data"]["name"] = corpus_type
 	obj_json["data"]["language"] = "English"
 	obj_json["data"]["bias"] = dict_aspect2bias
-	obj_json["data"]["output"] = path_comb_output
 
 	obj_json["model"]["name"] = model_name
-	obj_json["model"]["results"]["overall"]["error_case"] = errorCase_list
 	obj_json["model"]["results"]["overall"]["performance"] = holistic_performance
 	obj_json["model"]["results"]["overall"]["confidence_low"] = confidence_low
 	obj_json["model"]["results"]["overall"]["confidence_up"] = confidence_up
 	obj_json["model"]["results"]["fine_grained"] = dict_fineGrained
 
 
+
+
+	# add errorAnalysis -- holistic
+	obj_json["model"]["results"]["overall"]["error_case"] = errorCase_list
+
+
+
+	# for Calibration
 	ece = 0
 	dic_calibration = None
 	if is_print_ece:
@@ -429,10 +452,9 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 							   size_of_bin=10, dataset=corpus_type, model=model_name)
 
 	obj_json["model"]["results"]["calibration"] = dic_calibration
-	#print(dic_calibration)
 
 
-	save_json(obj_json, "./instantiate.json")
+
 	save_json(obj_json, fn_write_json)
