+++ explainaboard/tasks/pos/template.json
