+++ explainaboard/tasks/tc/tensoreval.py
@@ -4,8 +4,20 @@ import numpy
 import sys
 # sys.path.append("./src")
 # from src.utils import *
-#from errorAnalysis import *
+# from src.errorAnalysis import *
+
 from ..src.errorAnalysis import *
+
+
+
+
+
+# def tuple2str(triplet):
+# 	res = ""
+# 	for v in triplet:
+# 		res += str(v) + "_"
+# 	return res.rstrip("_")
+
 
 
 
@@ -18,7 +30,10 @@ def sent2list(sent):
 
 
 
-def getAspectValue(sample_list, dict_aspect_func):
+def getAspectValue(sent_list, sample_list_tag, sample_list_tag_pred, dict_aspect_func):
+
+
+
 
 
 	dict_span2aspectVal = {}
@@ -35,24 +50,21 @@ def getAspectValue(sample_list, dict_aspect_func):
 
 
 	sample_id = 0
-	for  info_list in sample_list:
+	for  sent, tag, tag_pred in zip(sent_list, sample_list_tag, sample_list_tag_pred):
+
+
+		dict_sid2sent[str(sample_id)] = format4json_tc(sent)
 
 
 
-		#
-		#
-		#
-		# word_list = wordSegment(sent).split(" ")
+		word_list = wordSegment(sent).split(" ")
 
-		# Sentence	Entities	Paragraph	True Relation Label	Predicted Relation Label
-		# Sentence Length	Paragraph Length	Number of Entities in Ground Truth Relation	Average Distance of Entities
 
-		sent, entities, paragraph, true_label, pred_label, sent_length, para_length, n_entity, avg_distance = info_list
 
-		dict_sid2sent[str(sample_id)] = format4json_tc(entities + "|||" + sent)
+		sent_length = len(word_list)
 
-		sent_pos = tuple2str((sample_id, true_label))
-		sent_pos_pred = tuple2str((sample_id, pred_label))
+		sent_pos = tuple2str((sample_id, tag))
+		sent_pos_pred = tuple2str((sample_id, tag_pred))
 
 		# Sentence Length: sentALen
 		aspect = "sLen"
@@ -60,31 +72,13 @@ def getAspectValue(sample_list, dict_aspect_func):
 			dict_span2aspectVal[aspect][sent_pos] = float(sent_length)
 			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(sent_length)
 
-		# Paragraph Length: pLen
-		aspect = "pLen"
-		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = float(para_length)
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(para_length)
-
-		# Number of Entity: nEnt
-		aspect = "nEnt"
-		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = float(n_entity)
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(n_entity)
-
-		# Average Distance: avgDist
-		aspect = "avgDist"
-		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = float(avg_distance)
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = float(avg_distance)
-
-
 
 		# Tag: tag
 		aspect = "tag"   ############## MUST Be Gold Tag for text classification task
 		if aspect in dict_aspect_func.keys():
-			dict_span2aspectVal[aspect][sent_pos] = true_label
-			dict_span2aspectVal_pred[aspect][sent_pos_pred] = true_label
+			dict_span2aspectVal[aspect][sent_pos] = tag
+			dict_span2aspectVal_pred[aspect][sent_pos_pred] = tag
+
 
 
 		sample_id += 1
@@ -93,6 +87,146 @@ def getAspectValue(sample_list, dict_aspect_func):
 	return  dict_span2aspectVal, dict_span2aspectVal_pred, dict_sid2sent
 
 
+def get_probability_right_or_not(file_path):
+    """
+
+    :param file_path: the file_path is the path to your file.
+
+    And the path must include file name.
+
+    the file name is in this format: test_dataset_model.tsv.
+
+    the file_path must in the format: /root/path/to/your/file/test_dataset.tsv
+
+    The file must in this format:
+    sentence\tground_truth\tpredict_label\tprobability\tright_or_not
+    if prediction is right, right_or_not is assigned to 1, otherwise 0.
+
+    """
+
+    import pandas as pd
+    import numpy as np
+
+    result = pd.read_csv(file_path, sep='\t', header=None)
+
+    probability_list = np.array(result[3]).tolist()
+    right_or_not_list = np.array(result[4]).tolist()
+
+    return probability_list, right_or_not_list
+
+
+def get_raw_list(probability_list, right_or_not_list):
+    total_raw_list = []
+
+    for index in range(len(right_or_not_list)):
+        total_raw_list.append([probability_list[index], right_or_not_list[index]])
+    return total_raw_list
+
+
+def calculate_ece(result_list):
+    ece = 0
+    size = 0
+    tem_list = []
+    for value in result_list:
+        if value[2] == 0:
+            tem_list.append(0)
+            continue
+        size = size + value[2]
+        error = abs(float(value[0]) - float(value[1]))
+        tem_list.append(error)
+
+    if size == 0:
+        return -1
+
+    for i in range(len(result_list)):
+        ece = ece + ((result_list[i][2] / size) * tem_list[i])
+
+    return ece
+
+
+def divide_into_bin(size_of_bin, raw_list):
+    bin_list = []
+    basic_width = 1 / size_of_bin
+
+    for i in range(0, size_of_bin):
+        bin_list.append([])
+
+    for value in raw_list:
+        probability = value[0]
+        isRight = value[1]
+        if probability == 1.0:
+            bin_list[size_of_bin - 1].append([probability, isRight])
+            continue
+        for i in range(0, size_of_bin):
+            if (probability >= i * basic_width) & (probability < (i + 1) * basic_width):
+                bin_list[i].append([probability, isRight])
+
+    result_list = []
+    for i in range(0, size_of_bin):
+        value = bin_list[i]
+        if len(value) == 0:
+            result_list.append([0, 0, 0])
+            continue
+        total_probability = 0
+        total_right = 0
+        for result in value:
+            total_probability = total_probability + result[0]
+            total_right = total_right + result[1]
+        result_list.append([total_probability / len(value), total_right / (len(value)), len(value)])
+
+    return result_list
+
+
+def process_all(file_path, size_of_bin=10, dataset='atis', model='lstm-self-attention'):
+    """
+
+    :param file_path: the file_path is the path to your file.
+
+    And the path must include file name.
+
+    the file name is in this format: test_dataset_model.tsv.
+
+    the file_path must in the format: /root/path/to/your/file/test_dataset.tsv
+
+    The file must in this format:
+    sentence\tground_truth\tpredict_label\tprobability\tright_or_not
+    if prediction is right, right_or_not is assigned to 1, otherwise 0.
+
+    :param size_of_bin: the numbers of how many bins
+
+    :param dataset: the name of the dataset
+
+    :param model: the name of the model
+
+    :return:
+    ece :the ece of this file
+    dic :the details of the ECE information in json format
+    """
+
+    # import json
+
+    probability_list, right_or_not_list = get_probability_right_or_not(file_path)
+
+    raw_list = get_raw_list(probability_list, right_or_not_list)
+
+    bin_list = divide_into_bin(size_of_bin, raw_list)
+
+    ece = calculate_ece(bin_list)
+    dic = OrderedDict()
+    dic['dataset-name'] = dataset
+    dic['model-name'] = model
+    dic['ECE'] = ece
+    dic['details'] = []
+    basic_width = 1 / size_of_bin
+    for i in range(len(bin_list)):
+        tem_dic = {}
+
+        bin_name = format(i * basic_width, '.2g') + '-' + format((i+1) * basic_width, '.2g')
+        tem_dic = {'interval':bin_name,'average_accuracy': bin_list[i][1], 'average_confidence': bin_list[i][0],
+                             'samples_number_in_this_bin': bin_list[i][2]}
+        dic['details'].append(tem_dic)
+
+    return ece, dic
 
 def evaluate(task_type = "ner", analysis_type = "single", systems = [], output = "./output.json", is_print_ci = False, is_print_case = False, is_print_ece = False):
 
@@ -106,10 +240,8 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 	corpus_type = "dataset_name"
 	model_name = "model_name"
 	path_preComputed = ""
-	path_aspect_conf = "./interpret_eval/tasks/re/conf.aspects"
-	path_json_input = "./interpret_eval/tasks/re/template.json"
-	# path_aspect_conf = "./tasks/re/conf.aspects"
-	# path_json_input = "./tasks/re/template.json"
+	path_aspect_conf = "./explainaboard/tasks/tc/conf.aspects"
+	path_json_input = "./explainaboard/tasks/tc/template.json"
 	fn_write_json = output
 
 
@@ -135,31 +267,38 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 			print("PreComputed directory:\t", dict_preComputed_path[aspect])
 
 
-	sample_list, sent_list, entity_list, true_list, pred_list = file_to_list_re(path_text)
+
+	sent_list, true_label_list, pred_label_list = file_to_list_tc(path_text)
 
 	errorCase_list = []
 	if is_print_case:
-		errorCase_list = getErrorCase_re(sent_list, entity_list, true_list, pred_list)
+		errorCase_list = getErrorCase_tc(sent_list, true_label_list, pred_label_list)
 		print(" -*-*-*- the number of error casse:\t", len(errorCase_list))
 
 
 
 
-	dict_span2aspectVal, dict_span2aspectVal_pred,  dict_sid2sent  = getAspectValue(sample_list, dict_aspect_func)
 
 
-	holistic_performance = accuracy(true_list, pred_list)
-	holistic_performance = format(holistic_performance, '.3g')
+
+
+
+
 
 
 	# Confidence Interval of Holistic Performance
 	confidence_low, confidence_up = 0,0
 	if is_print_ci:
-		confidence_low, confidence_up = compute_confidence_interval_acc(true_list, pred_list, n_times=1000)
+		confidence_low, confidence_up = compute_confidence_interval_acc(true_label_list, pred_label_list, n_times=1000)
 
 
 
-	dict_span2aspectVal, dict_span2aspectVal_pred,  dict_sid2sent  = getAspectValue(sample_list, dict_aspect_func)
+
+	dict_span2aspectVal, dict_span2aspectVal_pred,  dict_sid2sent  = getAspectValue(sent_list, true_label_list, pred_label_list, dict_aspect_func)
+
+
+	holistic_performance = accuracy(true_label_list, pred_label_list)
+	holistic_performance = format(holistic_performance, '.3g')
 
 
 
@@ -200,7 +339,7 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 		dict_bucket2span_pred[aspect] = bucketAttribute_SpecifiedBucketInterval(dict_span2aspectVal_pred[aspect],
 																				dict_bucket2span[aspect].keys())
 		# dict_bucket2span_pred[aspect] = __selectBucktingFunc(func[0], func[1], dict_span2aspectVal_pred[aspect])
-		dict_bucket2f1[aspect] = getBucketAcc_with_errorCase_re(dict_bucket2span[aspect], dict_bucket2span_pred[aspect], dict_sid2sent, is_print_ci, is_print_case)
+		dict_bucket2f1[aspect] = getBucketAcc_with_errorCase(dict_bucket2span[aspect], dict_bucket2span_pred[aspect], dict_sid2sent, is_print_ci, is_print_case)
 		aspect_names.append(aspect)
 	print("aspect_names: ", aspect_names)
 
@@ -256,12 +395,12 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 			#bucket_value = format(v[0]*100,'.4g')
 			bucket_value   = format(v[0], '.4g')
 			n_sample = v[1]
-			confidence_low_bucket = format(v[2], '.4g')
-			confidence_up_bucket  = format(v[3], '.4g')
+			confidence_low = format(v[2], '.4g')
+			confidence_up  = format(v[3], '.4g')
 			bucket_error_case = v[4]
 
 			# instantiation
-			dict_fineGrained[aspect].append({"bucket_name":bucket_name, "bucket_value":bucket_value, "num":n_sample, "confidence_low":confidence_low_bucket, "confidence_up":confidence_up_bucket, "bucket_error_case":bucket_error_case})
+			dict_fineGrained[aspect].append({"bucket_name":bucket_name, "bucket_value":bucket_value, "num":n_sample, "confidence_low":confidence_low, "confidence_up":confidence_up, "bucket_error_case":bucket_error_case})
 
 
 
@@ -293,66 +432,7 @@ def evaluate(task_type = "ner", analysis_type = "single", systems = [], output =
 	#print(dic_calibration)
 
 
-
+	save_json(obj_json, "./instantiate.json")
 	save_json(obj_json, fn_write_json)
 
 
-
-
-
-
-
-
-
-
-
-
-#
-# def main():
-#
-# 	parser = argparse.ArgumentParser(description='Interpretable Evaluation for NLP')
-#
-#
-# 	parser.add_argument('--task', type=str, required=True,
-# 						help="absa")
-#
-# 	parser.add_argument('--ci', type=str, required=False, default= False,
-# 						help="True|False")
-#
-# 	parser.add_argument('--case', type=str, required=False, default= False,
-# 						help="True|False")
-#
-# 	parser.add_argument('--ece', type=str, required=False, default= False,
-# 						help="True|False")
-#
-#
-# 	parser.add_argument('--type', type=str, required=False, default="single",
-# 						help="analysis type: single|pair|combine")
-# 	parser.add_argument('--systems', type=str, required=True,
-# 						help="the directories of system outputs. Multiple one should be separated by comma, for example, system1,system2 (no space)")
-#
-# 	parser.add_argument('--output', type=str, required=True,
-# 						help="analysis output file")
-# 	args = parser.parse_args()
-#
-#
-# 	is_print_ci = args.ci
-# 	is_print_case = args.case
-# 	is_print_ece = args.ece
-#
-# 	task = args.task
-# 	analysis_type = args.type
-# 	systems = args.systems.split(",")
-# 	output = args.output
-#
-#
-# 	print("task", task)
-# 	print("type", analysis_type)
-# 	print("systems", systems)
-# 	# sample_list = file_to_list_re(systems[0])
-# 	# print(sample_list[0])
-# 	evaluate(task_type=task, analysis_type=analysis_type, systems=systems, output=output, is_print_ci = is_print_ci, is_print_case = is_print_case, is_print_ece = is_print_ece)
-#
-# # python tensoreval.py  --task re --systems ./test_re.tsv --output ./a.json
-# if __name__ == '__main__':
-# 	main()
