--- interpret_eval/__init__.py
@@ -1,5 +1,0 @@
-__version__ = "0.1.7"
-
-import interpret_eval.interpret_eval_main
-import interpret_eval.tasks
-
